+++ iam-login-service/src/test/java/it/infn/mw/iam/test/api/account/multi_factor_authentication/authenticator_app/AuthenticatorAppSettingsControllerTests.java
@@ -19,10 +19,8 @@ import static it.infn.mw.iam.api.account.multi_factor_authentication.authenticat
 import static it.infn.mw.iam.api.account.multi_factor_authentication.authenticator_app.AuthenticatorAppSettingsController.DISABLE_URL;
 import static it.infn.mw.iam.api.account.multi_factor_authentication.authenticator_app.AuthenticatorAppSettingsController.ENABLE_URL;
 import static it.infn.mw.iam.api.account.multi_factor_authentication.authenticator_app.AuthenticatorAppSettingsController.MFA_SECRET_NOT_FOUND_MESSAGE;
-import static org.hamcrest.CoreMatchers.containsString;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
@@ -31,7 +29,6 @@ import static org.springframework.security.test.web.servlet.setup.SecurityMockMv
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.log;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import java.util.Optional;
@@ -49,9 +46,6 @@ import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.context.WebApplicationContext;
 import org.springframework.web.util.NestedServletException;
 
-import dev.samstevens.totp.exceptions.QrGenerationException;
-import dev.samstevens.totp.qr.QrData;
-import dev.samstevens.totp.qr.QrGenerator;
 import it.infn.mw.iam.api.account.multi_factor_authentication.IamTotpMfaService;
 import it.infn.mw.iam.config.mfa.IamTotpMfaProperties;
 import it.infn.mw.iam.core.user.exception.MfaSecretAlreadyBoundException;
@@ -64,7 +58,6 @@ import it.infn.mw.iam.test.TestUtils;
 import it.infn.mw.iam.test.multi_factor_authentication.MultiFactorTestSupport;
 import it.infn.mw.iam.test.util.WithAnonymousUser;
 import it.infn.mw.iam.test.util.WithMockMfaUser;
-import it.infn.mw.iam.test.util.WithMockOAuthUser;
 import it.infn.mw.iam.test.util.WithMockPreAuthenticatedUser;
 import it.infn.mw.iam.test.util.annotation.IamMockMvcIntegrationTest;
 import it.infn.mw.iam.util.mfa.IamTotpMfaEncryptionAndDecryptionUtil;
@@ -86,10 +79,6 @@ public class AuthenticatorAppSettingsControllerTests extends MultiFactorTestSupp
 
   @MockBean
   private IamTotpMfaProperties iamTotpMfaProperties;
-
-  @MockBean
-  private QrGenerator qrGenerator;
-
 
   @BeforeClass
   public static void init() {
@@ -122,29 +111,6 @@ public class AuthenticatorAppSettingsControllerTests extends MultiFactorTestSupp
     // TODO called twice for some reason?
     verify(accountRepository, times(2)).findByUsername(TEST_USERNAME);
     verify(totpMfaService, times(1)).addTotpMfaSecret(account);
-  }
-
-  @Test
-  @WithMockUser(username = TEST_USERNAME)
-  public void testAddSecretThrowsQrGenerationException() throws Exception {
-    IamAccount account = cloneAccount(TEST_ACCOUNT);
-    when(accountRepository.findByUsername(TEST_USERNAME)).thenReturn(Optional.of(account));
-
-    IamTotpMfa totpMfa = cloneTotpMfa(TOTP_MFA);
-    totpMfa.setSecret(IamTotpMfaEncryptionAndDecryptionUtil.encryptSecret(TOTP_MFA_SECRET,
-        iamTotpMfaProperties.getPasswordToEncryptOrDecrypt()));
-    when(totpMfaService.addTotpMfaSecret(account)).thenReturn(totpMfa);
-
-    when(qrGenerator.generate(any(QrData.class))).thenThrow(
-        new QrGenerationException("Simulated QR generation failure", new RuntimeException()));
-
-    mvc.perform(put(ADD_SECRET_URL))
-      .andExpect(status().isBadRequest())
-      .andExpect(content().string(containsString("Could not generate QR code")));
-
-    verify(accountRepository, times(2)).findByUsername(TEST_USERNAME);
-    verify(totpMfaService, times(1)).addTotpMfaSecret(account);
-    verify(qrGenerator, times(1)).generate(any(QrData.class));
   }
 
   @Test
@@ -202,28 +168,6 @@ public class AuthenticatorAppSettingsControllerTests extends MultiFactorTestSupp
   @Test
   @WithMockUser(username = TEST_USERNAME)
   public void testEnableAuthenticatorApp() throws Exception {
-    IamAccount account = cloneAccount(TEST_ACCOUNT);
-
-    IamTotpMfa totpMfa = cloneTotpMfa(TOTP_MFA);
-    totpMfa.setActive(true);
-    totpMfa.setAccount(account);
-    totpMfa.setSecret(IamTotpMfaEncryptionAndDecryptionUtil.encryptSecret(TOTP_MFA_SECRET,
-        iamTotpMfaProperties.getPasswordToEncryptOrDecrypt()));
-    String totp = "123456";
-
-    when(totpMfaService.verifyTotp(account, totp)).thenReturn(true);
-    when(totpMfaService.enableTotpMfa(account)).thenReturn(totpMfa);
-
-    mvc.perform(post(ENABLE_URL).param("code", totp)).andExpect(status().isOk());
-
-    verify(accountRepository, times(2)).findByUsername(TEST_USERNAME);
-    verify(totpMfaService, times(1)).verifyTotp(account, totp);
-    verify(totpMfaService, times(1)).enableTotpMfa(account);
-  }
-
-  @Test
-  @WithMockOAuthUser(user = TEST_USERNAME, authorities = "ROLE_USER")
-  public void testEnableAuthenticatorAppViaOauthAuthn() throws Exception {
     IamAccount account = cloneAccount(TEST_ACCOUNT);
 
     IamTotpMfa totpMfa = cloneTotpMfa(TOTP_MFA);
